Move validators to a single file
